fixed dependency in the UI within other classes
