+++ src/main/java/design/MazeDesignerOutputBoundary.java
@@ -1,10 +1,6 @@
 package design;
 
-import javafx.scene.control.Button;
-import javafx.stage.Stage;
-
 public interface MazeDesignerOutputBoundary {
-    public void updateMazeUI(Button[][] buttonarray);
+    public String[][] updateMazeUI(String[][] buttonarray);
 
-    public void outerWallEdit(Stage primaryStage);
 }
